support  passing the mode to router.go (#20)
## Summary

This PR enables the `mode` parameter for a routing call to be passed to `Router.go`. This is useful for consumers to be able to invoke the router in `pop` mode, which runs the handlers for the current URL without the side effect of updating the history stack.

This is exactly what happens when `popstate` is called, but it can be useful for other callers in a world where the history stack is altered from outside the context of the Router.